northd: Fix segfault when changing the router to a gateway router.
When a distributed router with a gateway router port is changed
to a gateway router (options:chassis=<ch>), during the recompute
run, op->cr_port is not created and instead a warning is logged
because of bad configuration.

The commit b52ba504ed6f is accessing op->cr_port even though cr_port
is not created for the 'op' and this results in a segfault.

This patch fixes it.

Fixes: b52ba504ed6f ("northd: Fix relying on naming coincidences.")
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>