northd: Remove the list of groups from mcast info.
The removal is needed for IGMP I-P node, because the ovn_datapath
shouldn't hold reference to struct that might have different lifetime
and might be freed during different node run. The same logic applies
in the opposite direction, the list holds pointer to list inside
ovn_datapath.

This also allows to better isolate logical flows that are related
to igmp in functions that can be reused for the IGMP handler.

Co-authored-by: Jacob Tanenbaum <jtanenba@redhat.com>
Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>
Suggested-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>