tests: Fix incorrect invocations of send_garp.
Some calls were missed during the test refactoring and that caused
packets to not be generated and not sent in the end.

Check is added to the receive command to make sure the packet sending
actually works.

Fixes: 0814c5b164a9 ("tests: Implement send_garp using scapy.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
(cherry picked from commit 85f274f4a8700581eb4b8404271f7f0e9104a938)