tests: Fix use of bash arrays in MAC binding tests.
This only works on systems where bash is the default shell, which is
not the case on many systems, including our CI.  So, the test was not
actually running in CI in full.  The IP addresses used in the third
check were also incorrect, so the test would fail if it was running.

While at it, removing the unnecessary extra line.

Fixes: 7abae8142841 ("northd: Fix an issue wrt mac binding aging.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
(cherry picked from commit 9ba3c44fcc15c8933b5e688212c9ee7985440ce8)