mac-cache: Fix MAC binding entry lookup for timestamp refresh.
MAC binding entry lookups are broken in two ways:

1. Port key is explicitly set to zero while creating entries from Sb
   records, but it is populated from the OpenFlow statistics.
2. Sb record pointer is populated while creating entries from Sb
   records, but it is not available while constructing MAC binding
   data from OpenFlow statistics replies.

Since both the port key and the Sb UUID are parts of the hash and the
equality check, it's actually impossible to find MAC binding cache
entries from OpenFlow statistics.

Fix that by not populating port keys from OpenFlow and using a cookie
instead of Sb UUID.

This set of information doesn't guarantee uniqueness, unfortunately,
but we can't actually get it if we don't have a full index to use.
To reduce chances of collision we could extend the cookie in the
future to 64 bits or store the full Sb UUID in the action notes in OF.
But it seems excessive at the moment, as we need to at least bring
the timestamp update to a working condition.  Having a cookie plus the
IP address and the datapath key should be a good enough distinguishing
factor.

The issue wasn't caught by the existing test because it waits for
timestamps to change without checking if they actually changing in
the right way.  There are two timestamps in the string and we wait
for the string to change and then checking that it's not empty.
What happens is that one entry expires and we're left with only one
timestamp, which is the same as before.  But the string changed and
it's not empty, so the check passes.

The printing function is split instead of just being modified. A
part of it will be used in the next commits, so splitting here to
avoid touching it twice.

Fixes: a68ac1a4a1ea ("mac-cache: Properly handle deletion of SB mac_bindings.")
Reported-at: https://issues.redhat.com/browse/FDP-1131
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>