controller: Omit alert for FDB and MAC binding timestamp.
The timestamp column is read by ovn-controller only in order to
implement aging.  However, ovn-controller doesn't need to react to
changes in the timestamp value so we can disable change tracking and
alerting for these columns.  That allows us to remove the helper
function for checking if the only updated column is the timestamp.
The same applies to both FDB and MAC binding.

Fixes: b57f60a617b4 ("controller: Add MAC cache I-P node")
Fixes: 6d4c23a6ea44 ("controller: Add FDB support to MAC cache I-P node")
Signed-off-by: Ales Musil <amusil@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>