northd: Add support for spine-leaf logical switch topology.
In L2 topology of OVN cluster, all the VIFs of the same network are
attached to the same logical switch.  Meaning, this switch has one LSP
for each container/VM in the cluster.  So, in a 10-node cluster with
100 containers/VMs per node, the switch will have 1000 ports.

This topology has limitations:

  - Processing of a very large switch may be expensive for northd as
    some changes to the switch may require reprocessing logical flows
    for all the ports.

  - OVN supports up to 32K ports in a single switch, i.e., a single
    network can't have more than 32K containers/VMs, which may be a
    serious limitation for large scale setups, as it can only
    accommodate a single /17 network.

A solution is to split this one large switch into a spine-leaf topology
by having a smaller switch per node that only have local LSPs and a
distributed switch that connects all these node switches together.
(Nothing prevents us from connecting multiple distributed switches,
but it's not the main focus of this work.)

Switches are connected via newly introduced "switch" type ports. These
ports have implicit "unknown" addresses, since the address information
is not leaked between directly connected switches.

Such topology should allow:

  - Better locality of changes: Adding new containers/VMs should only
    require adding LSPs to a single node switch on the same node where
    the container/VM is located.  And the switch updates should be
    cheaper for northd to process.

  - Higher number of containers/VMs in the cluster: We could create
    32K ports per node and not across the whole cluster.

  - Still have an L2 topology without any routers involved for EW
    traffic.

Unlike having node switches with localnet ports, there is no need
to have a dedicated provider network.  At the same time on OpenFlow
level it looks almost like if the spine switch didn't exist, which is
expected since it's a single L2 domain and the whole topology is
known to ovn-controller.  And the separation really exists only on
the logical level, providing a nice separation.

And while this topology may be useful as-is, it makes more sense in
the context of OVN interconnect, support for which will be added in
the next commits.

Implementation doesn't require changes in ovn-controller, we just
create a new 'switch' LSP type that translates into existing "patch"
port binding added into MC_unknown.

'ovn-nbctl lsp-set-type' command is updated to accept optional 'peer'
argument for switch ports, e.g.:

  ovn-nbctl lsp-set-type ls1-p1 switch peer=ls2-p2 \
         -- lsp-set-type ls2-p2 switch peer=ls1-p1

Adding a new 'lsp-set-peer' seemed excessive when it would only work
for a single port type.

Tests added to cover a local and a distributed spine-leaf topology.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>