controller: Update OVS submodule.
This bumps the OVS submodule to the current tip of branch-3.5,
8ceccbb404d5 ("github: Update fedora image to 41."), and brings in
needed changes so that the ovs submodule exports route table related
functionality.

It also picks up 63785d16db11 ("fedora: Don't generate
network-scripts-openvswitch for ver > 40.") which is needed for moving
OVN CI to Fedora 41 images.

As the ovs submodule has bumped its DPDK version we need to do so as
well.

Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>