northd: Fix broadcast of all traffic within a spine switch.
Currently, FDB learning is not enabled for the switch-switch ports
connecting switches in the Spine-Leaf topology.  This is causing a
traffic broadcast in the spine switch for every packet.  Even in cases
where it doesn't end up creating extra work in the datapath (since
ovn-controller knows the whole topology), this still creates a lot
of extra work for OpenFlow processing, since we need to evaluate
those rules for every connected switch during upcall processing.
And in cases where leaf switches have ports with unknown addresses,
we may end up unnecessarily broadcasting the actual traffic within
the datapath to those ports.

Fix that by enabling FDB learning for switch ports as it is already
done for other ports with unknown addresses.

Tests are enhanced to check that FDB is actually working and that
we're not unnecessarily broadcasting traffic.

For the case with interconnect this only partially solves the problem,
since we can't learn from remote ports, and so the packets are still
broadcasted to all the zones on the transit spine switch.  At least,
now the traffic will be dropped on the unrelated leaf switches, once
they learn that the actual destination is behind the spine switch from
which the packet just arrived.  Learning from remote ports to stop
the broadcasting will be addressed in the next commits.

Having an upcall per switch seems a little excessive, but it should
only happen once per MAC address and should not be a problem after
all the addresses are learned.  Also, with the main use case being
a transit switch, learning will only be triggered for switches local
to the availability zone, which should be a relatively small number.
However, this learning per switch behavior might still be a good
candidate for a future improvement.

Fixes: a2db2b2f263a ("northd: Add support for spine-leaf logical switch topology.")
Suggested-by: Numan Siddique <numans@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>
(cherry picked from commit d03d194407d6d1f1a70bf23ede9c911fe4deea4b)