ovn-nbctl.8: Document broadcasting behavior of 'switch' ports.
In a Spine-Leaf topology, initial packets for MAC addresses that
are not learned yet will be broadcasted within a spine switch.
It may be desirable for some configurations, because the nodes
don't need to be configured with all the addresses of all the ports
as with a single flat switch.  This improves control plane scalability,
provides locality of changes and allows to have more ports within a
single L2 network.  But the broadcast may also cause issues on the
dataplane side of things, so may not be desirable for other types
of setups or traffic patterns.

Document the broadcasting and learning behavior, so users can make
an informed choice while planning their virtual network topology.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>