controller: Update ovn-monitor-all documentation.
With the change in 354d766ca892 ("ovn-controller: Remove monitor all
of chassis private.") the documentation of the "ovn-monitor-all"
external_id became a bit stale.  The Chassis_Private table is always
conditionally monitored.

Update the documentation to reflect that.

While at it, also fix up some more minor issues in the documentation of
the flag.

Fixes: 3c355e35298b ("ovn-controller.c: Support option "ovn-monitor-all".")
Fixes: 354d766ca892 ("ovn-controller: Remove monitor all of chassis private.")
Reported-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>