actions: Make sure all action opcodes have string representation.
In order to print the action opcode name in the log there was a need
to add it into the macro inside "ovnact_op_to_string()". Make the
macro global one in the header instead, that allows the string
represantation to contain all opcodes because it is taken directly
from the definition in the header file.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
(cherry picked from commit e585fbd96407f35dd50f76a25b95942cc16254d9)