en-advertised-route-sync: Don't assume that switch peers are routers.
A peer of a switch can be another switch, so the port type has to be
checked.  The missing check doesn't seem to lead to crashes, but it
leads to addresses of switch-switch ports not being advertised.

Fix that by checking the port type instead of assuming.

Fixes: 93f541f342f9 ("northd: Allow announcing individual host routes.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
(cherry picked from commit 7cefb0f7c791ba92fb5152f8cf925eda379fd728)