controller: Avoid unnecessary recomputes due to bfd_chassis.
bfd_chassis node was recomputed when sb_global columns such as nb_cfg
were updated. If, in addition, SB was read only at that moment (i.e. an
transaction was in fly), then the update was cancelled and resulted
in a full recompute.

There is no dependency between sb_global and bfd_chassis, so remove the
handler.

This was highlighted as causing random failures in tests such as
"ACL with Port Group conjunction flow efficiency".

Fixes: ea619bf64ce1 ("controller: Add bfd_chassis engine.")
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>