Fix missing load balancer hairpin flows.
If we add a datapath local to the current hv to a load balancer which was only added to
non local datapaths, then hairpin flows were missing.

For instance, let's have sw0 and sw0p1 bound to hv0, and sw1 and sw1p1 bound to hv1
Let's also have a load balancer lb1 added to sw0.
When lb1 is added to sw1, hairpin flows were missing.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>