tests: Set inactivity_probe for ovn-remote.
Common macros set up a SSL or TCP connection for communication
between ovn-controller and SB DB in the test environment.

This connection is subject to the default inactivity_probe value
of 5000ms.

On slow systems, this may not be enough.  Consequently set the
inactivity_probe to 15000ms on both ends of the connection.

Reported-at: https://launchpad.net/bugs/2103444
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>