northd: Use next-hop network for SNAT when lb_force_snat_ip=router_ip.
When a gateway router has a load balancer configured, the option
lb_force_snat_ip=router_ip can be set so that OVN SNATs load balanced
packets to the logical router's egress interface's IP address, that is, the
port chosen as "outport" in the lr_in_ip_routing stage.

However, this was only designed to work when one network was configured
on the logical router outport. When multiple networks were configured,
OVN's behavior was to simply choose the lexicographically first IP address for
SNAT. This often lead to an incorrect address being used for SNAT.

To fix this, two main components have been added:
 1. A new flag, flags.network_id. It is 4 bits and stores an index.
 2. A new stage in the router ingress pipeline, lr_in_network_id.

Now in the stage lr_in_network_id, OVN generates flows that assign
flags.network_id with an index, which is chosen by looping through the
networks on the port, and selecting the one that contains IP of the
next hop. flags.network_id is then matched on later in lr_out_snat
and the network at that index will be chosen for SNAT.

However, if there are more than 16 networks, flags.network_id will be 0
for networks 17 and up. Then, for those networks, the first
lexicographical network will be chosen for SNAT.

There are also two lower priority flows: one with the old behavior to make
upgrades smooth, and another that sets flags.network_id to zero when the
next-hop doesn't belong to any of the port networks.

Three tests have been added to verify that:
 1. The correct network is chosen for SNAT.
 2. The new and updated flows with flags.network_id are correct.
 3. When there's one IPv6 addr and no IPv4 addr, the priority-105 flow is not
    added.

And tests that were broken by this new behavior have been updated.

Reported-at: https://issues.redhat.com/browse/FDP-871
Reported-at: https://mail.openvswitch.org/pipermail/ovs-dev/2024-October/417717.html
Signed-off-by: Rosemarie O'Riorden <rosemarie@redhat.com>
Acked-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>