controller: Added support for port mirroring in OVN overlay.
Added new mirror port type which is the link between the source
and the target port, its parent is the target port. VLAN tagging
is not required for mirror ports — packets are transmitted without
VLAN tag. Also since lport mirror works due to logical flows, we
don't need to pass information about the mirror to ovs.

Signed-off-by: Alexandra Rukomoinikova <arukomoinikova@k2.cloud>
Signed-off-by: Vladislav Odintsov <vlodintsov@k2.cloud>
Co-authored-by: Vladislav Odintsov <vlodintsov@k2.cloud>
Tested-by: Ivan Burnin <iburnin@k2.cloud>
Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>