inc-proc-eng: Update return type of run() callbacks.
An engine node's run() callback is intended to do two things:
1. Recalculate node data based on inputs.
2. Change the state of the node to indicate how the data was affected
   during the function call.

(1) is well-documented and clear based on the stated intent of the run()
callback from comments.

(2) is less well-defined and very easy to overlook. I, personally, have
lost hours trying to debug why input data was being read as NULL, only
to figure out that the input node's run() callback had forgotten to
update the node state to EN_UPDATED.

This change seeks to solve this problem by forcing the run() callback
to return the new engine node state to the incremental engine. The
engine internals are then responsible for updating the node state based
on the return value of the run() callback.

This patch is intended not to alter any existing logic in the code. It
should be exactly the same before and after this patch.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>