inc-proc-eng: Update return type of change handlers.
Change handlers' return type is a boolean: true if the handler could
handle the changed input data, and false if the handler could not handle
the input data. Unfortunately, a change handler also has a hidden task
it must do: update the node state if its data has been updated.

To make this responsibility more clear, this commit updates the return
type of change handlers to be an enum. This way, the change handler can
indicate if the change cannot be handled, if the change could be handled
but the node data is unchanged, or if the change could be handled and
the node data is updated. The incremental engine uses this return value
to update the node state.

This has a positive side effect as well, as this change now eliminates
the possibility of inputs overwriting state set by previous inputs.
Consider if node X has inputs A and B. A's change handler sets X's state
to EN_UPDATED. Then B's change handler sets X's state to EN_UNCHANGED. B
has overwritten the update that A made and therefore, X appears to be
unchanged, even though its data has been updated. With this commit, this
can no longer happen, since the incremental engine library itself is now
responsible for setting the state based on the change handler results.
The incremental engine will not change a node's state back to
EN_UNCHANGED if a previous change handler has set the state TO
EN_UPDATED.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>