northd: Find ovn_datapath from a single hmap.
The ovn_datapath_from_sbrec() function requires two hmaps to be passed
in. It needs the hmap of logical_switches and the hmap of
logical_routers.

In many cases, there is only a single northbound type that we care to
try to find. In this commit, we add a new version of
ovn_datapath_from_sbrec() that takes a single hmap.

Since there are several places in the code that passed NULL as one of
the hmap arguments to ovn_datapath_from_sbrec(), these have been
converted to use ovn_datapath_from_sbrec_() instead.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>