ci: Add permanent fix for /etc/hosts file.
This is OVN adoptation of OvS patch:

On two separate occasions GitHub added random garbage into the hosts
file breaking our tests.  This change adds a permanent workaround for
this kind of stuff.  It will remove everything that doesn't look like
a correct syntax from the file.

The regex is not perfect, but it should be sufficient for most cases.
It allows empty lines, comments and a valid 'IP NAME[ NAME]...' lines,
where 'IP' resembles an IP address and the 'NAME' consists of valid
DNS characters.  Under normal conditions the diff should always be
empty.

Co-authored-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Ales Musil <amusil@redhat.com>