binding: Avoid 100% CPU when postponing claims.
When a port claim was postponed (due to two gateways fighting to own a port),
and if the reason to claim the port was removed (e.g. bfd status has changed from
down to up on a low priority chassis), then ovn-controller was looping at 100% cpu
as the port stayed in the postponed list.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>